локаторы переехали в отдельные файлы
