+++ pages/main_page.py
@@ -1,40 +1,10 @@
-from selenium.webdriver.common.by import By
 from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.support import expected_conditions
 
+from locators.main_page import MainPageLocators as Locators
+
 
 class MainPage:
-    # Кнопка согласия использования кук
-    cookie_confirm_button = [By.ID, 'rcc-confirm-button']
-
-    # Кнопка Заказать в шапке
-    header_order_btn = [By.XPATH, './/div[@class="Header_Nav__AGCXC"]/button[text()="Заказать"]']
-
-    # Кнопка Заказать в подвале
-    footer_order_btn = [By.XPATH, './/div[@class="Home_FinishButton__1_cWm"]/button[text()="Заказать"]']
-
-    @staticmethod
-    def item_heading_text(x):
-        """
-        Вопрос в пункте разделе "Вопросы о важном"
-
-        :param x: номер пункта
-
-        :return: локатор на текст вопроса
-        """
-
-        return By.XPATH, f".//div[@id = 'accordion__heading-{x}']"
-
-    @staticmethod
-    def item_panel_text(x):
-        """
-        Ответ в пункте разделе "Вопросы о важном"
-
-        :param x: номер пункта
-
-        :return: локатор на текст ответа
-        """
-        return By.XPATH, f".//div[@id='accordion__heading-{x}']/following::div/p"
 
     def __init__(self, driver):
         self.driver = driver
@@ -46,13 +16,13 @@ class MainPage:
         :param item_number: Номер заголовка
         """
 
-        element = self.driver.find_element(*self.item_heading_text(item_number))
+        element = self.driver.find_element(*Locators.item_heading_text(item_number))
         self.driver.execute_script("arguments[0].scrollIntoView();", element)
 
         WebDriverWait(self.driver, 3).until(
-            expected_conditions.element_to_be_clickable(self.item_heading_text(item_number)))
+            expected_conditions.element_to_be_clickable(Locators.item_heading_text(item_number)))
 
-        self.driver.find_element(*self.item_heading_text(item_number)).click()
+        self.driver.find_element(*Locators.item_heading_text(item_number)).click()
 
     def accordion_item_texts(self, item_number):
         """
@@ -64,19 +34,19 @@ class MainPage:
         """
 
         WebDriverWait(self.driver, 3).until(
-            expected_conditions.element_to_be_clickable(self.item_panel_text(item_number)))
+            expected_conditions.element_to_be_clickable(Locators.item_panel_text(item_number)))
 
-        return (self.driver.find_element(*self.item_heading_text(item_number)).text,
-                self.driver.find_element(*self.item_panel_text(item_number)).text)
+        return (self.driver.find_element(*Locators.item_heading_text(item_number)).text,
+                self.driver.find_element(*Locators.item_panel_text(item_number)).text)
 
     def click_cookie_confirm_button(self):
         """
         Кликает по кнопке согласия с куками, если таковая найдена
         """
 
-        elements = self.driver.find_elements(*self.cookie_confirm_button)
+        elements = self.driver.find_elements(*Locators.cookie_confirm_button)
         if len(elements) > 0:
-            self.driver.find_element(*self.cookie_confirm_button).click()
+            self.driver.find_element(*Locators.cookie_confirm_button).click()
 
     def click_order_button(self, button):
         """
@@ -98,11 +68,11 @@ class MainPage:
         Кликает по кнопке Заказать в шапке
         """
 
-        self.driver.find_element(*self.header_order_btn).click()
+        self.driver.find_element(*Locators.header_order_btn).click()
 
     def click_footer_order_button(self):
         """
         Кликает по кнопке Заказать в подвале
         """
 
-        self.driver.find_element(*self.footer_order_btn).click()
+        self.driver.find_element(*Locators.footer_order_btn).click()
